requirements.txt
llama-index
llama-index-embeddings-huggingface
python-dotenv
groq
gradio
requests
tqdm
torch
transformers
sentence-transformers
numpy<2
